Renamed Basic Rule to Direct Rule for Tree Tent tests
