--- src/test/java/puzzles/treetent/rules/EmptyFieldBasicRuleTest.java
@@ -1,88 +1,0 @@
-package puzzles.treetent.rules;
-
-import edu.rpi.legup.model.tree.TreeNode;
-import edu.rpi.legup.model.tree.TreeTransition;
-import edu.rpi.legup.puzzle.treetent.TreeTent;
-import edu.rpi.legup.puzzle.treetent.TreeTentBoard;
-import edu.rpi.legup.puzzle.treetent.TreeTentCell;
-import edu.rpi.legup.puzzle.treetent.TreeTentType;
-import edu.rpi.legup.puzzle.treetent.rules.EmptyFieldDirectRule;
-import edu.rpi.legup.save.InvalidFileFormatException;
-import legup.MockGameBoardFacade;
-import legup.TestUtilities;
-import org.junit.Assert;
-import org.junit.BeforeClass;
-import org.junit.Test;
-
-import java.awt.*;
-
-public class EmptyFieldBasicRuleTest {
-
-    private static final EmptyFieldDirectRule RULE = new EmptyFieldDirectRule();
-    private static TreeTent treetent;
-
-    @BeforeClass
-    public static void setUp() {
-        MockGameBoardFacade.getInstance();
-        treetent = new TreeTent();
-    }
-
-    @Test
-    public void EmptyFieldTest() throws InvalidFileFormatException {
-        TestUtilities.importTestBoard("puzzles/treetent/rules/EmptyFieldBasicRule/EmptyField", treetent);
-        TreeNode rootNode = treetent.getTree().getRootNode();
-        TreeTransition transition = rootNode.getChildren().get(0);
-        transition.setRule(RULE);
-
-        TreeTentBoard board = (TreeTentBoard) transition.getBoard();
-
-        TreeTentCell cell1 = board.getCell(1, 1);
-        cell1.setData(TreeTentType.GRASS);
-
-        board.addModifiedData(cell1);
-
-        Assert.assertNull(RULE.checkRule(transition));
-
-        for (int i = 0; i < board.getHeight(); i++) {
-            for (int k = 0; k < board.getWidth(); k++) {
-                Point point = new Point(k, i);
-                if (point.equals(cell1.getLocation())) {
-                    Assert.assertNull(RULE.checkRuleAt(transition, board.getCell(k, i)));
-                }
-                else {
-                    Assert.assertNotNull(RULE.checkRuleAt(transition, board.getCell(k, i)));
-                }
-            }
-        }
-    }
-
-    @Test
-    public void DiagonalTreeTest() throws InvalidFileFormatException {
-        TestUtilities.importTestBoard("puzzles/treetent/rules/EmptyFieldBasicRule/DiagonalTree", treetent);
-        TreeNode rootNode = treetent.getTree().getRootNode();
-        TreeTransition transition = rootNode.getChildren().get(0);
-        transition.setRule(RULE);
-
-        TreeTentBoard board = (TreeTentBoard) transition.getBoard();
-
-        TreeTentCell cell1 = board.getCell(1, 1);
-        cell1.setData(TreeTentType.GRASS);
-
-        board.addModifiedData(cell1);
-
-        Assert.assertNull(RULE.checkRule(transition));
-
-        for (int i = 0; i < board.getHeight(); i++) {
-            for (int k = 0; k < board.getWidth(); k++) {
-                Point point = new Point(k, i);
-                if (point.equals(cell1.getLocation())) {
-                    Assert.assertNull(RULE.checkRuleAt(transition, board.getCell(k, i)));
-                }
-                else {
-                    Assert.assertNotNull(RULE.checkRuleAt(transition, board.getCell(k, i)));
-                }
-            }
-        }
-    }
-}
-
