+++ src/test/java/puzzles/treetent/rules/FinishWithGrassDirectRuleTest.java
@@ -6,7 +6,7 @@ import edu.rpi.legup.puzzle.treetent.TreeTent;
 import edu.rpi.legup.puzzle.treetent.TreeTentBoard;
 import edu.rpi.legup.puzzle.treetent.TreeTentCell;
 import edu.rpi.legup.puzzle.treetent.TreeTentType;
-import edu.rpi.legup.puzzle.treetent.rules.LastCampingSpotDirectRule;
+import edu.rpi.legup.puzzle.treetent.rules.FinishWithGrassDirectRule;
 import edu.rpi.legup.save.InvalidFileFormatException;
 import legup.MockGameBoardFacade;
 import legup.TestUtilities;
@@ -16,9 +16,9 @@ import org.junit.Test;
 
 import java.awt.*;
 
-public class LastCampingSpotBasicRuleTest {
+public class FinishWithGrassDirectRuleTest {
 
-    private static final LastCampingSpotDirectRule RULE = new LastCampingSpotDirectRule();
+    private static final FinishWithGrassDirectRule RULE = new FinishWithGrassDirectRule();
     private static TreeTent treetent;
 
     @BeforeClass
@@ -29,7 +29,7 @@ public class LastCampingSpotBasicRuleTest {
 
     @Test
     public void EmptyFieldTest() throws InvalidFileFormatException {
-        TestUtilities.importTestBoard("puzzles/treetent/rules/LastCampingSpotBasicRule/LastCampingSpot", treetent);
+        TestUtilities.importTestBoard("puzzles/treetent/rules/FinishWithGrassBasicRule/FinishWithGrass", treetent);
         TreeNode rootNode = treetent.getTree().getRootNode();
         TreeTransition transition = rootNode.getChildren().get(0);
         transition.setRule(RULE);
@@ -37,16 +37,19 @@ public class LastCampingSpotBasicRuleTest {
         TreeTentBoard board = (TreeTentBoard) transition.getBoard();
 
         TreeTentCell cell1 = board.getCell(1, 0);
-        cell1.setData(TreeTentType.TENT);
+        cell1.setData(TreeTentType.GRASS);
+        TreeTentCell cell2 = board.getCell(2, 0);
+        cell2.setData(TreeTentType.GRASS);
 
         board.addModifiedData(cell1);
+        board.addModifiedData(cell2);
 
         Assert.assertNull(RULE.checkRule(transition));
 
         for (int i = 0; i < board.getHeight(); i++) {
             for (int k = 0; k < board.getWidth(); k++) {
                 Point point = new Point(k, i);
-                if (point.equals(cell1.getLocation())) {
+                if (point.equals(cell1.getLocation()) || point.equals(cell2.getLocation())) {
                     Assert.assertNull(RULE.checkRuleAt(transition, board.getCell(k, i)));
                 }
                 else {
