+++ src/test/java/puzzles/treetent/rules/TentForTreeDirectRuleTest.java
@@ -1,0 +1,17 @@
+package puzzles.treetent.rules;
+
+// This feature is no longer supported
+public class TentForTreeDirectRuleTest {
+
+//    private static final TentForTreeBasicRule RULE = new TentForTreeBasicRule();
+//    private static TreeTent treetent;
+//
+//    @BeforeClass
+//    public static void setUp() {
+//        MockGameBoardFacade.getInstance();
+//        treetent = new TreeTent();
+//    }
+}
+
+
+
