--- src/test/java/puzzles/treetent/rules/TreeForTentBasicRuleTest.java
@@ -1,33 +1,0 @@
-package puzzles.treetent.rules;
-
-import edu.rpi.legup.model.tree.TreeNode;
-import edu.rpi.legup.model.tree.TreeTransition;
-import edu.rpi.legup.puzzle.treetent.TreeTent;
-import edu.rpi.legup.puzzle.treetent.TreeTentBoard;
-import edu.rpi.legup.puzzle.treetent.TreeTentCell;
-import edu.rpi.legup.puzzle.treetent.TreeTentType;
-import edu.rpi.legup.puzzle.treetent.rules.TreeForTentDirectRule;
-import edu.rpi.legup.save.InvalidFileFormatException;
-import legup.MockGameBoardFacade;
-import legup.TestUtilities;
-import org.junit.Assert;
-import org.junit.BeforeClass;
-import org.junit.Test;
-
-import java.awt.*;
-
-// This feature is no longer supported
-public class TreeForTentBasicRuleTest {
-
-//    private static final TreeForTentBasicRule RULE = new TreeForTentBasicRule();
-//    private static TreeTent treetent;
-
-//    @BeforeClass
-//    public static void setUp() {
-//        MockGameBoardFacade.getInstance();
-//        treetent = new TreeTent();
-//    }
-}
-
-
-
