+++ src/test/java/puzzles/treetent/rules/TreeForTentDirectRuleTest.java
@@ -1,0 +1,17 @@
+package puzzles.treetent.rules;
+
+// This feature is no longer supported
+public class TreeForTentDirectRuleTest {
+
+//    private static final TreeForTentBasicRule RULE = new TreeForTentBasicRule();
+//    private static TreeTent treetent;
+
+//    @BeforeClass
+//    public static void setUp() {
+//        MockGameBoardFacade.getInstance();
+//        treetent = new TreeTent();
+//    }
+}
+
+
+
